increase initial heap size to 8MiB
This allows for 32GiB of RAM while keeping boot times in the barely
noticable range (1-2 seconds on my machine).
Increasing it to 16MiB would allow for 64GiB of RAM, but increase
boot time to >3secs, which is noticable and I find annoying, so we'll
leave it at that until we need to support more.