remove drop impl for Process
The current drop impl was pointless.
It would remove the process from the process tree,
which is the structure that owns all processes.
However, the only way that processes are passed around
in the kernel is as `Arc<Process>`, so unless the process
tree doesn't own a process anymore, the process won't be
dropped anyways.